wms-ai-engine/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
pillow
numpy
opencv-python
sentencepiece

